Update climatedynamics_component.jl
An exception is included in the TATM function, since we follow the method in Anderson et al (2012) a3 (the damage exponent) is uniformly distributed between 1.8 and 2.2,, in the damage component TATM is elevated to a3 and a negative number to the power of a fractional number is a complex number which rises an error in a Monte Carlo simulation.  So we include a3 as a parameter, only for checking.